+++ src/grepros/plugins/auto/__init__.py
@@ -1,15 +1,14 @@
 # -*- coding: utf-8 -*-
 """
+Auto-loaded plugins.
+
 ------------------------------------------------------------------------------
 This file is part of grepros - grep for ROS bag files and live topics.
 Released under the BSD License.
 
 @author      Erki Suurjaak
-@created     31.10.2021
+@created     17.12.2021
 @modified    17.12.2021
 ------------------------------------------------------------------------------
 """
-## @namespace grepros
-__title__        = "grepros"
-__version__      = "0.3.6.dev4"
-__version_info__ = (0, 3, 6, "dev4")
+## @namespace grepros.plugins.auto