+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog for package grepros
 * add --reindex-if-unindexed option
 * fix error in Postgres output if a nested JSON value is NaN
 * fix skipping some messages in ROS1 bag for different types with identical hashes
+* ensure no conflicts from changed message types or identical type hashes
 
 0.3.5 (2021-12-14)
 -------------------